Task 04 (phase 6): extract :core:ui (Navigation, ImeState, viewmodel infra)
Move Navigation.kt (screen transitions), ImeState, the ViewModel factory infra
(ViewModelAssistedFactory/ViewModelFactory/ViewModelKey) and ViewModelFactory-
Container into a new Compose :core:ui module. Navigation.kt keeps its
com.timhome.core.ui package (no consumer import change); ImeState, viewmodel/*
and ViewModelFactoryContainer are repackaged and their imports rewritten across
:feature:{auth,home,device,settings}, which now depend on :core:ui.

After this, :core/src holds only the DI wiring + Application
(CoreComponent, AppModule, ModularizationApplication) - the thin core.

Builds green: assembleDebug + detektDebug + ktlintCheck.
