+++ core/ui/build.gradle.kts
@@ -1,0 +1,9 @@
+plugins {
+    id("timHome.android.library")
+    id("timHome.android.libraryCompose")
+    id("timHome.quality.convention.plugin")
+}
+
+android {
+    namespace = "com.timhome.core.ui"
+}
