+++ core/ui/src/main/java/com/timhome/core/ui/di/ViewModelFactoryContainer.kt
@@ -1,7 +1,7 @@
-package com.timhome.core.di
+package com.timhome.core.ui.di
 
 import androidx.compose.runtime.Stable
-import com.timhome.core.utils.viewmodel.ViewModelFactory
+import com.timhome.core.ui.viewmodel.ViewModelFactory
 import javax.inject.Inject
 
 @Stable
