+++ core/ui/src/main/java/com/timhome/core/ui/viewmodel/ViewModelFactory.kt
@@ -1,4 +1,4 @@
-package com.timhome.core.utils.viewmodel
+package com.timhome.core.ui.viewmodel
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
