+++ feature/auth/src/main/java/com/timhome/auth/di/AuthenticationComponent.kt
@@ -2,7 +2,7 @@ package com.timhome.auth.di
 
 import com.timhome.auth.ui.signin.di.SignInModule
 import com.timhome.core.di.CoreComponent
-import com.timhome.core.di.ViewModelFactoryContainer
+import com.timhome.core.ui.di.ViewModelFactoryContainer
 import com.timhome.core.common.FeatureScope
 import dagger.Component
