+++ feature/auth/src/main/java/com/timhome/auth/ui/signin/SignInViewModel.kt
@@ -6,7 +6,7 @@ import androidx.lifecycle.viewModelScope
 import com.timhome.auth.domain.IAuthorizationRepository
 import com.timhome.core.common.NavigationDispatcher
 import com.timhome.core.common.mvi.MviViewModel
-import com.timhome.core.utils.viewmodel.ViewModelAssistedFactory
+import com.timhome.core.ui.viewmodel.ViewModelAssistedFactory
 import com.google.firebase.analytics.FirebaseAnalytics
 import dagger.assisted.Assisted
 import dagger.assisted.AssistedFactory
