+++ feature/device/build.gradle.kts
@@ -10,6 +10,7 @@ android {
 dependencies {
     implementation(project(":core"))
     implementation(project(":core:common"))
+    implementation(project(":core:ui"))
     implementation(project(":core:designsystem"))
     implementation(project(":core:database"))
     implementation(libs.firebase.perf)
