+++ feature/device/src/main/java/com/timhome/device/ui/device/DeviceDetailViewModel.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.timhome.core.common.NavigationDispatcher
-import com.timhome.core.utils.viewmodel.ViewModelAssistedFactory
+import com.timhome.core.ui.viewmodel.ViewModelAssistedFactory
 import com.timhome.device.data.model.ModuleModel
 import com.timhome.device.domain.IDeviceRepository
 import com.timhome.device.domain.models.DeviceModel
