+++ feature/device/src/main/java/com/timhome/device/ui/listDevice/DeviceListModule.kt
@@ -1,8 +1,8 @@
 package com.timhome.device.ui.listDevice
 
 import androidx.lifecycle.ViewModel
-import com.timhome.core.utils.viewmodel.ViewModelAssistedFactory
-import com.timhome.core.utils.viewmodel.ViewModelKey
+import com.timhome.core.ui.viewmodel.ViewModelAssistedFactory
+import com.timhome.core.ui.viewmodel.ViewModelKey
 import com.timhome.device.data.repository.DeviceRepository
 import com.timhome.device.domain.IDeviceRepository
 import dagger.Binds
