+++ feature/device/src/main/java/com/timhome/device/ui/listDevice/DeviceListScreen.kt
@@ -14,7 +14,7 @@ import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.lifecycle.viewmodel.compose.viewModel
 import com.timhome.core.designsystem.theme.HomeTheme
 import com.timhome.core.common.trace
-import com.timhome.core.utils.viewmodel.ViewModelFactory
+import com.timhome.core.ui.viewmodel.ViewModelFactory
 import com.timhome.device.domain.models.DeviceModel
 import com.timhome.device.ui.listDevice.composables.DeviceListItem
