+++ feature/device/src/main/java/com/timhome/device/ui/listDevice/DeviceListViewModel.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.timhome.core.common.NavigationDispatcher
-import com.timhome.core.utils.viewmodel.ViewModelAssistedFactory
+import com.timhome.core.ui.viewmodel.ViewModelAssistedFactory
 import com.timhome.device.data.repository.DeviceRepository
 import com.timhome.device.domain.models.DeviceModel
 import com.google.firebase.analytics.FirebaseAnalytics
