+++ feature/home/src/main/java/com/timhome/home/di/HomeModule.kt
@@ -2,8 +2,8 @@ package com.timhome.home.di
 
 import androidx.lifecycle.ViewModel
 import com.timhome.base.BaseScope
-import com.timhome.core.utils.viewmodel.ViewModelAssistedFactory
-import com.timhome.core.utils.viewmodel.ViewModelKey
+import com.timhome.core.ui.viewmodel.ViewModelAssistedFactory
+import com.timhome.core.ui.viewmodel.ViewModelKey
 import com.timhome.home.data.api.WeatherApi
 import com.timhome.home.ui.HomeViewModel
 import dagger.Module
