+++ feature/home/src/main/java/com/timhome/home/ui/navigation/HomeRoute.kt
@@ -6,7 +6,7 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
 import com.timhome.base.DaggerBaseComponent
 import com.timhome.core.ModularizationApplication
-import com.timhome.core.di.ViewModelFactoryContainer
+import com.timhome.core.ui.di.ViewModelFactoryContainer
 import com.timhome.core.ui.slideIntoContainer
 import com.timhome.core.ui.slideOutOfContainer
 import com.timhome.home.di.DaggerHomeComponent
