+++ feature/settings/src/main/java/com/timhome/settings/di/SettingComponent.kt
@@ -1,7 +1,7 @@
 package com.timhome.settings.di
 
 import com.timhome.core.di.CoreComponent
-import com.timhome.core.di.ViewModelFactoryContainer
+import com.timhome.core.ui.di.ViewModelFactoryContainer
 import com.timhome.core.common.FeatureScope
 import com.timhome.settings.ui.SettingModule
 import dagger.Component
