+++ feature/settings/src/main/java/com/timhome/settings/ui/SettingModule.kt
@@ -1,8 +1,8 @@
 package com.timhome.settings.ui
 
 import androidx.lifecycle.ViewModel
-import com.timhome.core.utils.viewmodel.ViewModelAssistedFactory
-import com.timhome.core.utils.viewmodel.ViewModelKey
+import com.timhome.core.ui.viewmodel.ViewModelAssistedFactory
+import com.timhome.core.ui.viewmodel.ViewModelKey
 import dagger.Binds
 import dagger.Module
 import dagger.multibindings.IntoMap
