+++ settings.gradle.kts
@@ -30,6 +30,7 @@ include(":core:database")
 include(":core:datastore")
 include(":core:data")
 include(":core:designsystem")
+include(":core:ui")
 include(":benchmark")
 include(":mock")
 include(":base")
